requirements.txt
pillow
imagefilter
